// packages/secretnote-ui/src/.openapi-stubs/models/LocalObjectNode.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { LocalObject } from './LocalObject';

export type LocalObjectNode = {
  id: string;
  epoch: number;
  order?: number;
  kind?: 'local';
  data: LocalObject;
};